+++ src/main/codex/codex-config-mirror.test.ts
@@ -281,6 +281,21 @@ describe('syncSystemConfigIntoManagedCodexHome', () => {
     expect(existsSync(getSystemConfigPath())).toBe(false)
   })
 
+  it('preserves an existing runtime config when the system config is blank', () => {
+    // Why: a 0-byte config.toml is what a half-written or unhydrated
+    // cloud-synced home shows, not a deliberate "erase all my settings".
+    mkdirSync(join(userDataDir, 'codex-runtime-home', 'home'), { recursive: true })
+    const runtimeConfig = ['model = "runtime-model"', '', '[features]', 'hooks = true', ''].join(
+      '\n'
+    )
+    writeFileSync(getRuntimeConfigPath(), runtimeConfig, 'utf-8')
+    writeFileSync(getSystemConfigPath(), '', 'utf-8')
+
+    syncSystemConfigIntoManagedCodexHome()
+
+    expect(readFileSync(getRuntimeConfigPath(), 'utf-8')).toBe(runtimeConfig)
+  })
+
   it('mirrors system config updates while preserving runtime-owned trust sections', () => {
     mkdirSync(join(userDataDir, 'codex-runtime-home', 'home'), { recursive: true })
     writeFileSync(
